Updates to documentation
Team - FYI, URL for "Network security and containment" under "Align network segmentation with enterprise segmentation strategy" is returning 404 error message. Also, it would be great if we can align the headings with the bullet points. E.g. Following are the bullet points in the article and their sequence in the document.
 
1- Align network segmentation with overall strategy (Sequence # 02 in the document for description)
2- Centralize network management and security (Sequence # 01)
3- Build a network containment strategy (Sequence # 04, there is another topic at #03 which isn't mentioned in the bullet points)
4- Define an internet edge strategy